pyproject.toml
[project]
name = "local-rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.2",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-ollama>=1.0.1",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
